Stable Build 04/04/2017

+++ Capstone Project/Assets/Scripts/Camera/Screen Shakes/ScreenShakeRequest.cs
@@ -6,7 +6,7 @@ public abstract class ScreenShakeRequest : ScriptableObject {
     [SerializeField]
     protected float _shakeAmount = 0.0f;
 
-    public abstract Vector3 Shake(byte key = 0);
+    public abstract Vector3 Shake(int key = 0);
 
     public virtual void ShakeRequest() {
